fix: schema migration
