+++ package.json
@@ -15,10 +15,12 @@
     "@neondatabase/serverless": "^0.6.0",
     "drizzle-orm": "^0.28.6",
     "next": "14.0.1",
+    "postgres": "^3.4.3",
     "react": "^18",
     "react-dom": "^18"
   },
   "devDependencies": {
+    "@faker-js/faker": "^8.2.0",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
